requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
mysql-connector-python
xlsxwriter